Lab4/pzpi-22-5-oleshchenko-mykyta-lab4/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.3
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
flasgger==0.9.5
flask-cors==3.0.10
python-dotenv==1.0.0
Werkzeug==2.0.3
